Change host IP from 127.0.0.1 to 0.0.0.0
This allows for connecting locally/remotely from other machines. By default 127.0.0.1 only allows connections from that very machine. Since this is not documented as such, I think this is a better default config since it's up to external configurations to allow other styles of connection (remote connections).

I didn't catch this at first and so I was wondering why I couldn't connect at all from my other machine.

If this was meant for security reasons, then you might want to mention that one will need to change it to 0.0.0.0 in order for other machines to locally/remotely connect to it.